+++ src/Libraries/SmartStore.Core/Domain/Security/PermissionRecord.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 
 namespace SmartStore.Core.Domain.Security
 {
@@ -12,6 +13,7 @@ namespace SmartStore.Core.Domain.Security
         /// <summary>
         /// Gets or sets the permission system name.
         /// </summary>
+        [Index]
         public string SystemName { get; set; }
         
         /// <summary>