+++ src/Presentation/SmartStore.Web/Models/Catalog/ManufacturerModel.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using SmartStore.Services.Localization;
+﻿using SmartStore.Services.Localization;
 using SmartStore.Services.Search;
 using SmartStore.Web.Framework.Modelling;
 using SmartStore.Web.Models.Media;
@@ -16,6 +15,7 @@ namespace SmartStore.Web.Models.Catalog
 
         public LocalizedValue<string> Name { get; set; }
         public LocalizedValue<string> Description { get; set; }
+        public LocalizedValue<string> BottomDescription { get; set; }
         public LocalizedValue<string> MetaKeywords { get; set; }
         public LocalizedValue<string> MetaDescription { get; set; }
         public LocalizedValue<string> MetaTitle { get; set; }