+++ changelog.md
@@ -25,6 +25,7 @@
 * #1733 Add ability to hide products from catalog, but not search.
 * #1754 Implement hidden boolean setting that controls which catalog search engine implementation should be used in backend.
 * #1776 Enable images and color values for search filters to be stored for specification attribute options.
+* #1793 Add field for manufacturer bottom description like for categories.
 
 ### Improvements
 * #1663 Make MeasureDimension and MeasureWeight localizable.